Add DOING 核心功能组件扩展设计

+++ unicorn-storage/unicorn-storage-api/src/main/java/org/rhine/unicorn/storage/api/Store.java
@@ -2,7 +2,7 @@ package org.rhine.unicorn.storage.api;
 
 public interface Store {
 
-    int write(Message message) throws WriteExcetion;
+    long write(Message message) throws WriteExcetion;
 
     Message randomAccess(Message message) throws ReadException;
