+++ unicorn-storage/unicorn-storage-redis/src/main/java/org/rhine/unicorn/storage/redis/RedisStore.java
@@ -8,7 +8,7 @@ import org.rhine.unicorn.storage.api.WriteExcetion;
 public class RedisStore implements Store {
 
     @Override
-    public int write(Message message) throws WriteExcetion {
+    public long write(Message message) throws WriteExcetion {
         return 0;
     }
